Remove names from stop edit
