+++ ui/src/components/stop-registry/stops/stop-details/basic-details/basic-details-form/index.ts
@@ -1,7 +1,5 @@
 export * from './schema';
-export * from './StopAbbreviationsFormRow';
 export * from './StopBasicDetailsForm';
-export * from './StopLabelAndNameFormRow';
-export * from './StopLongNameAndLocationFormRow';
+export * from './StopLabelAndLocationFormRow';
 export * from './StopOtherDetailsFormRow';
 export * from './StopTypesFormRow';
